+++ src/components/main/MainHeader.css
@@ -1,4 +1,4 @@
-.mainpage-header {
+.mainpage-appguideheader {
   position: absolute;
   width: 99.95%;
   top: 0;
@@ -12,14 +12,44 @@
   color: var(--color-white);
   font-family: var(--font-noto-sans-kr);
 }
-.mainpage-headertitle {
+.mainpage-headermyinfobutton1 {
+  position: absolute;
+  width: 14rem;
+  cursor: pointer;
+  border: 0;
+  padding: 0;
+  background-color: var(--light);
+  top: 2.344rem;
+  right: 3.875rem;
+  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
+  border-radius: var(--br-8xs);
+  height: 6.5rem;
+}
+.mainpage-headermyinfobutton1:hover {
+  background-color: #CEE6B9;
+}
+.mainpage-div57 {
+  position: absolute;
+  width: 14rem;
+  top: 15.87%;
+  right: 0;
+  font-size: var(--font-size-29xl);
+  font-family: var(--font-noto-sans-kr);
+  color: var(--color-black);
+  text-align: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.mainpage-headertitle2 {
   position: absolute;
   top: 2.125rem;
   left: calc(50% - 332.5px);
   width: 41.563rem;
   height: 6.938rem;
+  cursor: pointer;
 }
-.mainpage-headertitleimage-icon {
+.mainpage-headertitleimage-icon4 {
   position: absolute;
   height: 100%;
   top: 0;
@@ -29,7 +59,7 @@
   width: 7.5rem;
   object-fit: cover;
 }
-.mainpage-div6 {
+.mainpage-div58 {
   position: absolute;
   top: 0.9%;
   left: calc(50% - 223.5px);
@@ -38,4 +68,33 @@
   justify-content: center;
   width: 34.75rem;
   color: black;
+}
+.mainpage-headerbackbutton5 {
+  position: absolute;
+  width: 14rem;
+  cursor: pointer;
+  border: 0;
+  padding: 0;
+  background-color: var(--light);
+  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
+  border-radius: var(--br-8xs);
+  top: 2.344rem;
+  left: 3.875rem;
+  height: 6.5rem;
+}
+.mainpage-headerbackbutton5:hover {
+  background-color: #CEE6B9;
+}
+.mainpage-div59 {
+  position: absolute;
+  width: 14rem;
+  top: 16.83%;
+  left: 0;
+  font-size: var(--font-size-29xl);
+  font-family: var(--font-noto-sans-kr);
+  color: var(--color-black);
+  text-align: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
