+++ src/pages/main/AppGuide.jsx
@@ -2,6 +2,7 @@ import { useCallback, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useChatBot } from "../../context/ChatBotContext"; // useChatBot 임포트
 import ChatBot from "../../components/ChatBot"; // ChatBot 컴포넌트 임포트
+import MainHeader from "../../components/main/MainHeader";
 import MainFooter from "../../components/main/MainFooter";
 import "./AppGuide.css";
 
@@ -9,10 +10,6 @@ const AppGuide = () => {
   const navigate = useNavigate();
   const [selectedCategory, setSelectedCategory] = useState("");
   const { isChatBotActive , activateChatBot, chatBotStyle} = useChatBot(); // 챗봇 상태 가져오기
-
-  const onLogoContainerClick = useCallback(() => {
-    navigate("/home");
-  }, [navigate]);
 
   const onNeverBandClick = useCallback(() => {
     navigate("/maincontents");
@@ -71,14 +68,6 @@ const AppGuide = () => {
 
   const onButton3Click = useCallback(() => {
     navigate("/info");
-  }, [navigate]);
-
-  const onHeaderMyinfoButtonClick = useCallback(() => {
-    navigate("/mypagemyaccount");
-  }, [navigate]);
-
-  const onHeaderBackButtonClick = useCallback(() => {
-    navigate("/home");
   }, [navigate]);
 
   const onStartButtonClick = () => {
@@ -219,25 +208,7 @@ const AppGuide = () => {
         {isChatBotActive && (<div className="mainpage-chatbot-container"><ChatBot /></div>)}
       </section>
       </div>
-      <header className="mainpage-appguideheader">
-        <button
-          className="mainpage-headermyinfobutton1"
-          onClick={onHeaderMyinfoButtonClick}
-        >
-          <div className="mainpage-div57">내 정보</div>
-        </button>
-        <div className="mainpage-headertitle2" onClick={onLogoContainerClick}>
-          <img
-            className="mainpage-headertitleimage-icon4"
-            alt=""
-            src="main/image-1@2x.png"
-          />
-          <div className="mainpage-div58">디지털지름길</div>
-        </div>
-        <button className="mainpage-headerbackbutton5" onClick={onHeaderBackButtonClick}>
-          <div className="mainpage-div59">뒤로가기</div>
-        </button>
-      </header>
+      <MainHeader />
       <MainFooter />
     </div>
   );
