+++ src/pages/main/Community.jsx
@@ -2,6 +2,7 @@ import { useCallback,useState } from "react";
 import { useNavigate } from "react-router-dom";
 import {useChatBot} from "../../context/ChatBotContext"; //useChatBot context
 import ChatBot from "../../components/ChatBot" //ChatBot Component
+import MainHeader from "../../components/main/MainHeader";
 import MainFooter from "../../components/main/MainFooter";
 import "./Community.css";
 
@@ -9,10 +10,6 @@ const Community = () => {
   const navigate = useNavigate();
   const [selectedCategory, setSelectedCategory] = useState("");
   const {isChatBotActive, activateChatBot, chatBotStyle} = useChatBot(); //ChatBot Statement
-
-  const onLogoContainerClick = useCallback(() => {
-    navigate("/home");
-    }, [navigate]);
 
   const onBandButtonClick = useCallback(() => {
     setSelectedCategory("bando")
@@ -39,14 +36,6 @@ const Community = () => {
 
   const onButton3Click = useCallback(() => {
     navigate("/info");
-  }, [navigate]);
-
-  const onHeaderMyinfoButtonClick = useCallback(() => {
-    navigate("/mypagemyaccount");
-  }, [navigate]);
-
-  const onHeaderBackButtonClick = useCallback(() => {
-    navigate("/home");
   }, [navigate]);
 
   const onStartButtonClick = () => {
@@ -130,28 +119,7 @@ const Community = () => {
         {isChatBotActive && (<div className="mainpage-chatbot-container"><ChatBot /></div>)}
       </section>
       </main>
-      <header className="mainpage-communityheader1">
-        <button
-          className="mainpage-headermyinfobutton10"
-          onClick={onHeaderMyinfoButtonClick}
-        >
-          <div className="mainpage-div203">내 정보</div>
-        </button>
-        <div className="mainpage-headertitle11" onClick={onLogoContainerClick}>
-          <img
-            className="mainpage-headertitleimage-icon13"
-            alt=""
-            src="main/image-1@2x.png"
-          />
-          <div className="mainpage-div204">디지털지름길</div>
-        </div>
-        <button
-          className="mainpage-headerbackbutton14"
-          onClick={onHeaderBackButtonClick}
-        >
-          <div className="mainpage-div205">뒤로가기</div>
-        </button>
-      </header>
+      <MainHeader />
       <MainFooter />
     </div>
   );
