+++ src/pages/main/DeviceGuide.css
@@ -382,97 +382,6 @@
   color: var(--color-black);
   font-family: var(--font-noto-sans-kr);
 }
-.mainpage-div192 {
-  position: absolute;
-  top: 15.87%;
-  right: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 14rem;
-}
-.mainpage-headermyinfobutton9 {
-  position: absolute;
-  top: 2.344rem;
-  right: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  background-color: var(--light);
-  width: 14rem;
-  height: 6.5rem;
-  cursor: pointer;
-}
-.mainpage-headermyinfobutton9:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headertitleimage-icon12 {
-  position: absolute;
-  height: 100%;
-  top: 0;
-  bottom: 0;
-  left: calc(50% - 332.5px);
-  max-height: 100%;
-  width: 7.5rem;
-  object-fit: cover;
-}
-.mainpage-div193 {
-  position: absolute;
-  top: 0.9%;
-  left: calc(50% - 223.5px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 34.75rem;
-  color: black;
-}
-.mainpage-headertitle10 {
-  position: absolute;
-  top: 2.125rem;
-  left: calc(50% - 332.5px);
-  width: 41.563rem;
-  height: 6.938rem;
-  font-size: var(--font-size-56xl);
-  color: var(--color-white);
-  cursor: pointer;
-}
-.mainpage-div194,
-.mainpage-headerbackbutton13 {
-  position: absolute;
-  width: 14rem;
-}
-.mainpage-div194 {
-  top: 16.83%;
-  left: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-headerbackbutton13 {
-  top: 2.344rem;
-  left: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  background-color: var(--light);
-  height: 6.5rem;
-  cursor: pointer;
-}
-.mainpage-headerbackbutton13:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-appguideheader1 {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 11.188rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-29xl);
-  color: var(--color-black);
-  font-family: var(--font-noto-sans-kr);
-}
 .mainpage-deviceguide {
   width: 100%;
   position: relative;
