+++ src/pages/main/DeviceSubPage.css
@@ -232,76 +232,6 @@
 .mainpage-choose::-webkit-scrollbar {
   display: none;
 }
-.mainpage-headertitleimage-icon2 {
-  position: absolute;
-  height: 74.12%;
-  top: 14.12%;
-  bottom: 11.76%;
-  left: calc(50% - 254.5px);
-  max-height: 100%;
-  width: 5.188rem;
-  object-fit: cover;
-  cursor: pointer;
-}
-.mainpage-div18,
-.mainpage-div19 {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-div18 {
-  height: 72.94%;
-  top: 10.59%;
-  left: calc(50% - 959.5px);
-  font-size: var(--font-size-29xl);
-  color: black;
-  width: 119.375rem;
-  cursor: pointer;
-}
-.mainpage-div19 {
-  height: 67.33%;
-  top: 16.67%;
-  left: 0;
-  width: 14rem;
-}
-.mainpage-headerbackbutton1,
-.mainpage-headerbackbutton2 {
-  position: absolute;
-  top: 0.75rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  background-color: var(--light);
-  width: 14rem;
-  height: 3.75rem;
-  cursor: pointer;
-}
-.mainpage-headerbackbutton2:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headerbackbutton1:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headerbackbutton1 {
-  left: 2.813rem;
-}
-.mainpage-headerbackbutton2 {
-  right: 2.75rem;
-}
-.mainpage-maincontentsheader {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 5.313rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-17xl);
-  color: var(--color-black);
-  font-family: var(--font-noto-sans-kr);
-}
 .mainpage-copyright2 {
   position: absolute;
   top: 68.67%;
