+++ src/pages/main/Home.css
@@ -364,107 +364,6 @@
   padding-bottom: 10px;
   overflow-y: auto;
 }
-.mainpage-div116 {
-  position: absolute;
-  top: 15.87%;
-  right: 0;
-  font-size: var(--font-size-29xl);
-  font-family: var(--font-noto-sans-kr);
-  color: var(--color-black);
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 14rem;
-}
-.mainpage-headermyinfobutton5 {
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--light);
-  position: absolute;
-  top: 2.344rem;
-  right: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  width: 14rem;
-  height: 6.5rem;
-}
-.mainpage-headermyinfobutton5:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headertitleimage-icon8 {
-  position: absolute;
-  height: 100%;
-  top: 0;
-  bottom: 0;
-  left: calc(50% - 332.5px);
-  max-height: 100%;
-  width: 7.5rem;
-  object-fit: cover;
-}
-.mainpage-div117 {
-  position: absolute;
-  top: 0.9%;
-  left: calc(50% - 223.5px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 34.75rem;
-  cursor: default;
-  color: black;
-}
-.mainpage-headertitle6 {
-  position: absolute;
-  top: 2.125rem;
-  left: calc(50% - 332.5px);
-  width: 41.563rem;
-  height: 6.938rem;
-}
-.mainpage-div118,
-.mainpage-headerbackbutton9 {
-  position: absolute;
-  width: 14rem;
-  border-radius: 5px;
-}
-.mainpage-div118 {
-  top: 16.83%;
-  left: 0;
-  font-size: var(--font-size-29xl);
-  font-family: var(--font-noto-sans-kr);
-  color: var(--color-black);
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-headerbackbutton9 {
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--light);
-  top: 2.344rem;
-  left: 3.875rem;
-  border-radius: var(--br-8xs);
-  height: 6.5rem;
-}
-.mainpage-headerbackbutton9:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-homeheader {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 11.188rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-56xl);
-  color: var(--color-white);
-  font-family: var(--font-noto-sans-kr);
-}
 .mainpage-home {
   width: 100%;
   position: relative;
@@ -659,89 +558,6 @@
   display: flex;
   align-items: center;
   justify-content: center;
-}
-
-.mainpage-homeheader {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 11.188rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-56xl);
-  color: var(--color-white);
-  font-family: var(--font-noto-sans-kr);
-}
-.mainpage-headertitle6 {
-  position: absolute;
-  top: 2.125rem;
-  left: calc(50% - 332.5px);
-  width: 41.563rem;
-  height: 6.938rem;
-}
-.mainpage-headertitleimage-icon8 {
-  position: absolute;
-  height: 100%;
-  top: 0;
-  bottom: 0;
-  left: calc(50% - 332.5px);
-  max-height: 100%;
-  width: 7.5rem;
-  object-fit: cover;
-}
-.mainpage-div117 {
-  position: absolute;
-  top: 0.9%;
-  left: calc(50% - 223.5px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 34.75rem;
-  cursor: default;
-}
-.mainpage-headermyinfobutton5,
-.mainpage-headerbackbutton9 {
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--light);
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  width: 14rem;
-  height: 6.5rem;
-}
-.mainpage-headermyinfobutton5 {
-  position: absolute;
-  top: 2.344rem;
-  right: 3.875rem;
-}
-.mainpage-headerbackbutton9 {
-  position: absolute;
-  top: 2.344rem;
-  left: 3.875rem;
-}
-.mainpage-div116,
-.mainpage-div118 {
-  position: absolute;
-  width: 14rem;
-  font-size: var(--font-size-29xl);
-  font-family: var(--font-noto-sans-kr);
-  color: var(--color-black);
-  text-align: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-div116 {
-  top: 15.87%;
-  right: 0;
-}
-.mainpage-div118 {
-  top: 16.83%;
-  left: 0;
 }
 
 /* 챗봇 컨테이너 스타일 추가 */
