+++ src/pages/main/Home.jsx
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { useChatBot } from "../../context/ChatBotContext";
 import ChatBot from "../../components/ChatBot";
+import MainHeader from "../../components/main/MainHeader";
 import MainFooter from "../../components/main/MainFooter";
 import "./Home.css";
 
@@ -45,14 +46,6 @@ const Home = () => {
   const onLogoutButtonClick = useCallback(() => {
     setUsername("");
     navigate("/home");
-  }, [navigate]);
-
-  const onHeaderMyinfoButtonClick = useCallback(() => {
-    navigate("/mypagemyaccount");
-  }, [navigate]);
-
-  const onHeaderBackButtonClick = useCallback(() => {
-    navigate("/");
   }, [navigate]);
 
   const onStartButtonClick = () => {
@@ -103,7 +96,7 @@ const Home = () => {
               alt=""
               src="main/homemaincommunitybuttonimage@2x.png"
             />
-            <div className="mainpage-div108">커뮤니티</div>
+              <div className="mainpage-div108">커뮤니티</div>
           </button>
           <button
             className="mainpage-homemainguidebutton"
@@ -191,25 +184,7 @@ const Home = () => {
           )}
         </div>
       </main>
-      <header className="mainpage-homeheader">
-        <button
-          className="mainpage-headermyinfobutton5"
-          onClick={onHeaderMyinfoButtonClick}
-        >
-          <div className="mainpage-div116">내 정보</div>
-        </button>
-        <div className="mainpage-headertitle6">
-          <img
-            className="mainpage-headertitleimage-icon8"
-            alt=""
-            src="main/image-1@2x.png"
-          />
-          <div className="mainpage-div117">디지털지름길</div>
-        </div>
-        <button className="mainpage-headerbackbutton9" onClick={onHeaderBackButtonClick}>
-          <div className="mainpage-div118">뒤로가기</div>
-        </button>
-      </header>
+      <MainHeader />
       <MainFooter />
     </div>
   );
