+++ src/pages/main/MainContents.css
@@ -386,82 +386,6 @@
   color: var(--color-black);
   font-family: var(--font-noto-sans-kr);
 }
-.mainpage-headertitleimage-icon3 {
-  position: absolute;
-  height: 74.12%;
-  top: 14.12%;
-  bottom: 11.76%;
-  left: calc(50% - 254.5px);
-  max-height: 100%;
-  width: 5.188rem;
-  object-fit: cover;
-  cursor: pointer;
-}
-.mainpage-div36,
-.mainpage-div37 {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-div36 {
-  height: 72.94%;
-  top: 10.59%;
-  left: calc(50% - 959.5px);
-  width: 119.375rem;
-  cursor: pointer;
-  font-size: var(--font-size-29xl);
-  color: black;
-}
-.mainpage-div37 {
-  height: 67.33%;
-  top: 16.67%;
-  left: 0;
-  font-size: var(--font-size-17xl);
-  font-family: var(--font-noto-sans-kr);
-  color: var(--color-black);
-  text-align: center;
-  width: 14rem;
-}
-.mainpage-headerbackbutton3,
-.mainpage-headerbackbutton4 {
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--light);
-  position: absolute;
-  top: 0.75rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  width: 14rem;
-  height: 3.75rem;
-}
-.mainpage-headerbackbutton3 {
-  left: 2.813rem;
-}
-.mainpage-headerbackbutton3:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headerbackbutton4 {
-  right: 2.75rem;
-}
-.mainpage-headerbackbutton4:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-maincontentsheader1 {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 5.313rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-41xl);
-  color: var(--color-white);
-  font-family: var(--font-noto-sans-kr);
-}
 .mainpage-maincontents {
   width: 100%;
   position: relative;
