+++ src/pages/main/MyPageStorageBox.css
@@ -322,97 +322,6 @@
   color: var(--color-black);
   font-family: var(--font-noto-sans-kr);
 }
-.mainpage-div153 {
-  position: absolute;
-  top: 15.87%;
-  right: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 14rem;
-}
-.mainpage-headermyinfobutton7 {
-  position: absolute;
-  top: 2.344rem;
-  right: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  background-color: var(--light);
-  border-radius: var(--br-8xs);
-  width: 14rem;
-  height: 6.5rem;
-  cursor: pointer;
-}
-.mainpage-headermyinfobutton7:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headertitleimage-icon10 {
-  position: absolute;
-  height: 100%;
-  top: 0;
-  bottom: 0;
-  left: calc(50% - 332.5px);
-  max-height: 100%;
-  width: 7.5rem;
-  object-fit: cover;
-}
-.mainpage-div154 {
-  position: absolute;
-  top: 0.9%;
-  left: calc(50% - 223.5px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 34.75rem;
-  color: black;
-}
-.mainpage-headertitle8 {
-  position: absolute;
-  top: 2.125rem;
-  left: calc(50% - 332.5px);
-  width: 41.563rem;
-  height: 6.938rem;
-  font-size: var(--font-size-56xl);
-  color: var(--color-white);
-  cursor: pointer;
-}
-.mainpage-div155,
-.mainpage-headerbackbutton11 {
-  position: absolute;
-  width: 14rem;
-}
-.mainpage-div155 {
-  top: 16.83%;
-  left: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-headerbackbutton11 {
-  top: 2.344rem;
-  left: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  background-color: var(--light);
-  border-radius: var(--br-8xs);
-  height: 6.5rem;
-  cursor: pointer;
-}
-.mainpage-headerbackbutton11:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-mypageheader1 {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 11.188rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-29xl);
-  color: var(--color-black);
-  font-family: var(--font-noto-sans-kr);
-}
 .mainpage-copyright11 {
   position: absolute;
   top: 68.67%;
