+++ src/pages/main/ScrapingPage.css
@@ -154,98 +154,6 @@
   left: 3.49%;
   height: 50rem;
 }
-.mainpage-div60,
-.mainpage-headermyinfobutton2 {
-  position: absolute;
-  width: 14rem;
-}
-.mainpage-div60 {
-  top: 15.87%;
-  right: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-headermyinfobutton2 {
-  top: 2.344rem;
-  right: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  background-color: var(--light);
-  height: 6.5rem;
-  cursor: pointer;
-}
-.mainpage-headermyinfobutton2:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headertitleimage-icon5 {
-  position: absolute;
-  height: 100%;
-  top: 0;
-  bottom: 0;
-  left: calc(50% - 332.5px);
-  max-height: 100%;
-  width: 7.5rem;
-  object-fit: cover;
-}
-.mainpage-div61 {
-  position: absolute;
-  top: 0.9%;
-  left: calc(50% - 223.5px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 34.75rem;
-  color: black;
-}
-.mainpage-headertitle3 {
-  position: absolute;
-  top: 2.125rem;
-  left: calc(50% - 332.5px);
-  width: 41.563rem;
-  height: 6.938rem;
-  font-size: var(--font-size-56xl);
-  color: var(--color-white);
-  cursor: pointer;
-}
-.mainpage-div62,
-.mainpage-headerbackbutton6 {
-  position: absolute;
-  width: 14rem;
-}
-.mainpage-div62 {
-  top: 16.83%;
-  left: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.mainpage-headerbackbutton6 {
-  top: 2.344rem;
-  left: 3.875rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  background-color: var(--light);
-  height: 6.5rem;
-  cursor: pointer;
-}
-.mainpage-headerbackbutton6:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-scrapingheader {
-  position: absolute;
-  width: 99.95%;
-  top: 0;
-  right: 0.05%;
-  left: 0;
-  background-color: var(--main);
-  height: 11.188rem;
-  overflow: hidden;
-  text-align: center;
-  font-size: var(--font-size-29xl);
-  color: var(--color-black);
-  font-family: var(--font-noto-sans-kr);
-}
 .mainpage-copyright5 {
   position: absolute;
   top: 68.67%;
