+++ src/pages/main/ScrapingPage.jsx
@@ -2,6 +2,7 @@ import { useCallback, useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import {useChatBot} from "../../context/ChatBotContext"; // useChatBot context import
 import ChatBot from "../../components/ChatBot"; //ChatBot component import
+import MainHeader from "../../components/main/MainHeader";
 import MainFooter from "../../components/main/MainFooter";
 import "./ScrapingPage.css";
 import Career from "../Information/Career";
@@ -15,18 +16,7 @@ const ScrapingPage = () => {
   const location = useLocation();
   const [selectedCategory, setSelectedCategory] = useState(location.state?.selectedCategory);
   const { isChatBotActive , activateChatBot, chatBotStyle} = useChatBot(); // chatbot functions
-
-  const onLogoContainerClick = useCallback(() => {
-    navigate("/home");
-    }, [navigate]);
   
-  const onHeaderMyinfoButtonContainerClick = useCallback(() => {
-    navigate("/mypagemyaccount");
-  }, [navigate]);
-
-  const onHeaderBackButtonContainerClick = useCallback(() => {
-    navigate("/home");
-  }, [navigate]);
 //카테고리 변경 함수
 const handleCategoryChange = (category) => {
   setSelectedCategory(category);
@@ -82,28 +72,6 @@ switch (selectedCategory) {
           <button className={`mainpage-button16 ${selectedCategory === "culture" ? "selected":""}`} onClick={()=>handleCategoryChange("culture")}>문화</button>
         </div>
       </main>
-      <header className="mainpage-scrapingheader">
-        <div
-          className="mainpage-headermyinfobutton2"
-          onClick={onHeaderMyinfoButtonContainerClick}
-        >
-          <div className="mainpage-div60">내 정보</div>
-        </div>
-        <div className="mainpage-headertitle3" onClick={onLogoContainerClick}>
-          <img
-            className="mainpage-headertitleimage-icon5"
-            alt=""
-            src="main/image-1@2x.png"
-          />
-          <div className="mainpage-div61">디지털지름길</div>
-        </div>
-        <div
-          className="mainpage-headerbackbutton6"
-          onClick={onHeaderBackButtonContainerClick}
-        >
-          <div className="mainpage-div62">뒤로가기</div>
-        </div>
-      </header>
       <section className="mainpage-chatbot4" >
         <img className="mainpage-box-icon4" alt="" src="main/box.svg" />
         <div className="mainpage-intro5">
@@ -120,6 +88,7 @@ switch (selectedCategory) {
         <img className="mainpage-character-icon4" alt="" src="main/character@2x.png" />
         {isChatBotActive && (<div className="mainpage-chatbot-container"><ChatBot /></div>)}
       </section>
+      <MainHeader />
       <MainFooter />
     </div>
   );
