+++ src/pages/main/Sns.css
@@ -84,118 +84,6 @@
   left: calc(50% - 893px);
   height: 49.688rem;
 }
-.mainpage-div7,
-.mainpage-headermyinfobutton {
-  position: absolute;
-  width: 14rem;
-}
-.mainpage-div7 {
-  height: 67.33%;
-  top: 0.2rem;
-  left: 0;
-  font-size: var(--font-size-17xl);
-  font-family: var(--font-noto-sans-kr);
-  color: var(--color-black);
-  text-align: center;
-  width: 14rem;
-}
-.mainpage-headermyinfobutton {
-  right: 2.75rem;
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--light);
-  position: absolute;
-  top: 0.75rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  width: 14rem;
-  height: 3.75rem;
-}
-.mainpage-headermyinfobutton:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-headertitleimage-icon1 {
-  position: absolute;
-  height: 74.12%;
-  top: 14.12%;
-  bottom: 11.76%;
-  left: calc(50% - 254.5px);
-  max-height: 100%;
-  width: 5.188rem;
-  object-fit: cover;
-  cursor: pointer;
-}
-.mainpage-div8 {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 72.94%;
-  top: 10.59%;
-  left: calc(50% - 959.5px);
-  width: 119.375rem;
-  cursor: pointer;
-  font-size: var(--font-size-29xl);
-  color: black;
-}
-.mainpage-headertitle1 {
-  height: 72.94%;
-  top: 10.59%;
-  left: calc(50% - 959.5px);
-  width: 119.375rem;
-  cursor: pointer;
-  font-size: var(--font-size-29xl);
-}
-.mainpage-div9,
-.mainpage-headerbackbutton {
-  position: absolute;
-  width: 14rem;
-}
-.mainpage-div9 {
-  position: absolute;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 67.33%;
-  top: 16.67%;
-  left: 0;
-  font-size: var(--font-size-17xl);
-  font-family: var(--font-noto-sans-kr);
-  color: var(--color-black);
-  text-align: center;
-  width: 14rem;
-}
-.mainpage-headerbackbutton {
-  left: 2.813rem;
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  background-color: var(--light);
-  position: absolute;
-  top: 0.75rem;
-  box-shadow: 6px 6px 4px rgba(0, 0, 0, 0.25);
-  border-radius: var(--br-8xs);
-  width: 14rem;
-  height: 3.75rem;
-}
-.mainpage-headerbackbutton:hover {
-  background-color: #CEE6B9;
-}
-.mainpage-communityheader {
-    position: absolute;
-    width: 99.95%;
-    top: 0;
-    right: 0.05%;
-    left: 0;
-    background-color: var(--main);
-    height: 5.313rem;
-    overflow: hidden;
-    text-align: center;
-    font-size: var(--font-size-41xl);
-    color: var(--color-white);
-    font-family: var(--font-noto-sans-kr);
-}
 .mainpage-copyright1 {
   position: absolute;
   top: 68.67%;
